feat : 상품 수정, 상품 목록 조회 기능 추가

+++ src/main/java/org/example/cafe/domain/product/repository/ProductRepository.java
@@ -3,6 +3,6 @@ package org.example.cafe.domain.product.repository;
 import org.example.cafe.domain.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
 }
