+++ src/main/java/org/example/cafe/domain/product/service/ProductService.java
@@ -24,7 +24,7 @@ public class ProductService {
         return productRepository.save(product);
     }
 
-    public Optional<Product> findById(int id) {
+    public Optional<Product> findById(Long id) {
         return productRepository.findById(id);
     }
